requirements.txt
2captcha-python==1.5.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aioquic==1.2.0
aiosignal==1.3.1
anyio==3.7.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
attrs==24.2.0
beautifulsoup4==4.12.3
blinker==1.9.0
Brotli==1.1.0
bs4==0.0.2
certifi==2024.8.30
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.0
chromedriver-autoinstaller==0.6.4
chromedriver-binary==133.0.6855.0.0
click==8.1.8
cryptography==43.0.3
Flask==3.1.0
frozenlist==1.5.0
greenlet==3.1.1
h11==0.12.0
h2==4.1.0
hpack==4.0.0
hstspreload==2024.10.1
httpcore==0.15.0
httpx==0.25.1
hyperframe==6.0.1
idna==2.10
itsdangerous==2.2.0
Jinja2==3.1.5
kaitaistruct==0.10
ldap3==2.9.1
lxml==5.3.1
MarkupSafe==3.0.2
mitmproxy-macos==0.11.5
mitmproxy_rs==0.11.5
msgpack==1.1.0
multidict==6.1.0
outcome==1.3.0.post0
packaging==24.1
passlib==1.7.4
pipdeptree==2.24.0
playwright==1.48.0
propcache==0.2.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
publicsuffix2==2.20191221
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyee==12.0.0
pylsqpack==0.3.19
pyOpenSSL==24.2.1
pyparsing==3.2.0
pyperclip==1.9.0
PySocks==1.7.1
pyTelegramBotAPI==4.25.0
python-dotenv==1.0.1
PyVirtualDisplay==3.0
requests==2.32.3
rfc3986==1.5.0
ruamel.yaml==0.18.10
selenium==4.25.0
selenium-wire==5.1.0
service-identity==24.2.0
setuptools==75.8.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
telebot==0.0.5
tornado==6.4.2
trio==0.27.0
trio-websocket==0.11.1
TwoCaptcha==0.0.1
typing_extensions==4.12.2
undetected-chromedriver==3.5.5
urllib3==2.2.3
urwid==2.6.16
wcwidth==0.2.13
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.17.1
zstandard==0.23.0
